+++ src/dactyl_keyboard/cad/body/main.clj
@@ -12,209 +12,112 @@
 ;;; which point it will exist only as a parameter default, or disappear.
 
 (ns dactyl-keyboard.cad.body.main
-  (:require [scad-clj.model :as model]
-            [scad-tarmi.core :refer [mean]]
-            [scad-tarmi.maybe :as maybe]
-            [scad-klupe.iso :as threaded]
-            [scad-tarmi.util :refer [loft]]
-            [dactyl-keyboard.cad.misc :as misc]
+  (:require [dactyl-keyboard.cad.misc :as misc]
             [dactyl-keyboard.cad.place :as place]
-            [dactyl-keyboard.cad.key :as key]
-            [dactyl-keyboard.compass :as compass :refer [sharp-left sharp-right]]
-            [dactyl-keyboard.param.access :as access :refer [most-specific compensator]]))
+            [dactyl-keyboard.compass :as compass]
+            [dactyl-keyboard.param.access :as access :refer [compensator]]
+            [scad-clj.model :as model]
+            [scad-klupe.iso :as threaded]
+            [scad-tarmi.maybe :as maybe]))
 
 
 ;;;;;;;;;;;;;;;;;;
 ;; Rear Housing ;;
 ;;;;;;;;;;;;;;;;;;
 
-(defn rhousing-post [getopt]
-  (let [xy (getopt :main-body :rear-housing :wall-thickness)]
-    (model/cube xy xy (getopt :main-body :rear-housing :roof-thickness))))
-
-(defn- rhousing-height
-  "The precise height of (the center of) each top-level rhousing-post."
-  [getopt]
-  (- (getopt :main-body :rear-housing :height)
-     (/ (getopt :main-body :rear-housing :roof-thickness) 2)))
-
 (defn rhousing-properties
   "Derive characteristics from parameters for the rear housing."
   [getopt]
-  (let [cluster (getopt :main-body :rear-housing :position :cluster)
-        key-style (fn [coord] (most-specific getopt [:key-style] cluster coord))
-        row (last (getopt :key-clusters :derived :by-cluster cluster :row-range))
-        coords (getopt :key-clusters :derived :by-cluster cluster
-                       :coordinates-by-row row)
-        pairs (into [] (for [coord coords, side [:NNW :NNE]] [coord side]))
-        getpos (fn [[coord side]]
-                 (place/cluster-place getopt cluster coord
-                   (place/mount-corner-offset getopt (key-style coord) side)))
-        y-max (apply max (map #(second (getpos %)) pairs))
-        getoffset (partial getopt :main-body :rear-housing :position :offsets)
-        y-roof-s (+ y-max (getoffset :south))
-        y-roof-n (+ y-roof-s (getoffset :north))
-        z (rhousing-height getopt)
-        roof-sw [(- (first (getpos (first pairs))) (getoffset :west)) y-roof-s z]
-        roof-se [(+ (first (getpos (last pairs))) (getoffset :east)) y-roof-s z]
-        roof-nw [(first roof-sw) y-roof-n z]
-        roof-ne [(first roof-se) y-roof-n z]
-        between (fn [a b] (mapv #(/ % 2) (mapv + a b)))]
-   {:coordinate-corner-pairs pairs
-    ;; [x y z] coordinates on the topmost part of the roof:
-    :side {:N (between roof-nw roof-ne)
-           :NE roof-ne
-           :E (between roof-ne roof-se)
-           :SE roof-se
-           :S (between roof-se roof-sw)
-           :SW roof-sw
-           :W (between roof-sw roof-nw)
-           :NW roof-nw}
-    :end-coord {:W (first coords), :E (last coords)}}))
+  (let [prop (partial getopt :main-body :rear-housing)
+        [_ _ z :as ext-size] (prop :size)
+        ext-pos (assoc (place/offset-from-anchor getopt (prop :anchoring) 2)
+                       2 (/ z 2))
+        int-pos (update ext-pos 2 #(- % (/ (prop :thickness :roof) 2)))
+        ww (* 2 (prop :thickness :walls))
+        int-size (mapv - ext-size [ww ww (prop :thickness :roof)])
 
-(defn- rhousing-roof
-  "A cuboid shape between the four corners of the rear housing’s roof."
-  [getopt]
-  (let [get-side (partial getopt :main-body :rear-housing :derived :side)]
-    (apply model/hull
-      (map #(maybe/translate (get-side %) (rhousing-post getopt))
-           [:NW :NE :SE :SW]))))
+        front (prop :bevel :exterior)]
+    {:position {:exterior ext-pos
+                :interior int-pos
+                ;; Remaining positions are just offset from the preceding bases.
+                :hollow [0 (/ (second int-size) -2) 0]
+                :mask [0 (/ front 2) 0]}
+     :size {:exterior ext-size
+            :interior int-size
+            :hollow (update int-size 1 (partial * 2))
+            :mask (update ext-size 1 #(- % front))}
+     :mount-width (* 2.2 (prop :fasteners :bolt-properties :m-diameter))}))
 
-(defn rhousing-pillar-functions
-  "Make functions that determine the exact positions of rear housing walls.
-  This is an awkward combination of reckoning functions for building the
-  bottom plate in 2D and placement functions for building the case walls in
-  3D. Because they’re specialized, the ultimate return values are disturbingly
-  different."
-  ;; TODO: Refactor this along the lines of the central housing.
-  [getopt]
-  (let [cluster (getopt :main-body :rear-housing :position :cluster)
-        cluster-pillar
-          (fn [cardinal rhousing-turning-fn cluster-turning-fn]
-            ;; Make a function for a part of the key cluster wall.
-            (fn [reckon upper]
-              (let [coord (getopt :main-body :rear-housing :derived :end-coord cardinal)
-                    subject (if reckon [0 0 0] (key/web-post getopt))
-                    ;; For reckoning, return a 3D coordinate vector.
-                    ;; For building, return a sequence of web posts.
-                    picker (if reckon #(first (take-last 2 %)) identity)]
-                (picker
-                  (place/wall-edge-sequence getopt cluster upper
-                    [coord cardinal rhousing-turning-fn] subject)))))
-        rhousing-pillar
-          (fn [side]
-            ;; Make a function for a part of the rear housing.
-            ;; For reckoning, return a 3D coordinate vector.
-            ;; For building, return a hull of housing cubes.
-            {:pre [(compass/intermediates side)]}
-            (fn [reckon upper]
-              (let [subject (if reckon
-                              (place/rhousing-vertex-offset getopt side)
-                              (rhousing-post getopt))]
-                (apply (if reckon mean model/hull)
-                  (map #(place/rhousing-place getopt side % subject)
-                       (if upper [0 1] [1]))))))]
-    [(cluster-pillar :W sharp-right sharp-left)
-     (rhousing-pillar :WSW)
-     (rhousing-pillar :WNW)
-     (rhousing-pillar :NNW)
-     (rhousing-pillar :NNE)
-     (rhousing-pillar :ENE)
-     (rhousing-pillar :ESE)
-     (cluster-pillar :E sharp-left sharp-right)]))
-
-(defn- rhousing-wall-shape-level
-  "The west, north and east walls of the rear housing with connections to the
-  ordinary case wall."
-  [getopt is-upper-level joiner]
-  (loft
-    (reduce
-      (fn [coll function] (conj coll (joiner (function false is-upper-level))))
-      []
-      (rhousing-pillar-functions getopt))))
-
-(defn- rhousing-outer-wall
-  "The complete walls of the rear housing: Vertical walls and a bevelled upper
-  level that meets the roof."
-  [getopt]
-  (model/union
-    (rhousing-wall-shape-level getopt true identity)
-    (rhousing-wall-shape-level getopt false misc/bottom-hull)))
-
-(defn- rhousing-web
-  "An extension of a key cluster’s webbing onto the roof of the rear housing."
-  [getopt]
-  (let [cluster (getopt :main-body :rear-housing :position :cluster)
-        key-style (fn [coord] (most-specific getopt [:key-style] cluster coord))
-        pos-corner (fn [coord side]
-                     (place/cluster-place getopt cluster coord
-                       (place/mount-corner-offset getopt (key-style coord) side)))
-        sw (getopt :main-body :rear-housing :derived :side :SW)
-        se (getopt :main-body :rear-housing :derived :side :SE)
-        x (fn [coord side]
-            (max (first sw)
-                 (min (first (pos-corner coord side))
-                      (first se))))
-        y (second sw)
-        z (rhousing-height getopt)]
-   (loft
-     (reduce
-       (fn [coll [coord side]]
-         (conj coll
-           (model/hull
-             (place/cluster-place getopt cluster coord
-               (key/mount-corner-post getopt (key-style coord) side))
-             (model/translate [(x coord side) y z]
-               (rhousing-post getopt)))))
-       []
-       (getopt :main-body :rear-housing :derived :coordinate-corner-pairs)))))
-
-(defn- rhousing-mount-place [getopt side shape]
+(defn- place-mount [getopt side shape]
   {:pre [(compass/cardinals side)]}
-  (let [d (getopt :main-body :rear-housing :fasteners :bolt-properties :m-diameter)
-        offset (getopt :main-body :rear-housing :fasteners
-                 (side compass/short-to-long) :offset)
-        n (getopt :main-body :rear-housing :position :offsets :north)
-        t (getopt :main-body :rear-housing :roof-thickness)
-        h (threaded/datum d :hex-nut-height)
-        [sign base] (case side
-                      :W [+ (getopt :main-body :rear-housing :derived :side :SW)]
-                      :E [- (getopt :main-body :rear-housing :derived :side :SE)])
-        near (mapv + base [(+ (- (sign offset)) (sign d)) d (/ (+ t h) -2)])
-        far (mapv + near [0 (- n d d) 0])]
-   (model/hull
-     (model/translate near shape)
-     (model/translate far shape))))
+  (let [prop (partial getopt :main-body :rear-housing)
+        d (prop :fasteners :bolt-properties :m-diameter)
+        w (prop :derived :mount-width)
+        leeway (/ (- (prop :derived :size :exterior 1) w) 2)
+        position (fn [y]
+                   (place/rhousing-place getopt :interior side 0 nil
+                     [(+ (prop :fasteners (side compass/short-to-long) :offset)
+                         (* (compass/delta-x (compass/reverse side)) (/ w 2)))
+                      y
+                      (/ (threaded/datum d :hex-nut-height) -2)]))]
+   (apply model/hull
+     (for [op [- +]] (model/translate (position (op leeway)) shape)))))
 
 (defn- rhousing-mount-positive [getopt side]
   {:pre [(compass/cardinals side)]}
   (let [d (getopt :main-body :rear-housing :fasteners :bolt-properties :m-diameter)
-        w (* 2.2 d)]
-   (rhousing-mount-place getopt side
+        w (getopt :main-body :rear-housing :derived :mount-width)]
+   (place-mount getopt side
      (model/cube w w (threaded/datum d :hex-nut-height)))))
 
 (defn- rhousing-mount-negative [getopt side]
   {:pre [(compass/cardinals side)]}
   (let [d (getopt :main-body :rear-housing :fasteners :bolt-properties :m-diameter)]
    (model/union
-     (rhousing-mount-place getopt side
+     (place-mount getopt side
        (model/cylinder (/ d 2) 20))
      (if (getopt :main-body :rear-housing :fasteners :bosses)
-       (rhousing-mount-place getopt side
-         (threaded/nut {:m-diameter d :compensator (compensator getopt) :negative true}))))))
+       (place-mount getopt side
+         (threaded/nut {:m-diameter d
+                        :compensator (compensator getopt) :negative true}))))))
 
-(defn rear-housing
+(defn rear-housing-exterior
+  "A single polyhedron in place.
+  Exposed for use in shaping a bottom plate under the rear housing."
+  [getopt]
+  (place/rhousing-place getopt :exterior nil nil nil
+    (misc/bevelled-cuboid
+      (getopt :main-body :rear-housing :derived :size :exterior)
+      (getopt :main-body :rear-housing :bevel :exterior))))
+
+(defn rear-housing-model
   "A squarish box at the far end of a key cluster."
   [getopt]
-  (let [prop (partial getopt :main-body :rear-housing :fasteners)
+  (let [prop (partial getopt :main-body :rear-housing)
+        fast (partial prop :fasteners)
         pair (fn [function]
-               (model/union
-                 (when (prop :west :include) (function getopt :W))
-                 (when (prop :east :include) (function getopt :E))))]
-   (model/difference
-     (model/union
-       (rhousing-roof getopt)
-       (rhousing-web getopt)
-       (rhousing-outer-wall getopt)
-       (if (prop :bosses) (pair rhousing-mount-positive)))
-     (pair rhousing-mount-negative))))
+               (maybe/union
+                 (when (fast :west :include) (function getopt :W))
+                 (when (fast :east :include) (function getopt :E))))]
+    (model/difference
+      (model/union
+        (model/intersection
+          ;; The mask.
+          (place/rhousing-place getopt :exterior nil nil
+            (prop :derived :position :mask)
+            (apply model/cube (prop :derived :size :mask)))
+          ;; The main part of the housing:
+          ;; An extra deep interior subtracted from an exterior.
+          ;; Using the nominal interior model would usually leave four walls.
+          (model/difference
+            (rear-housing-exterior getopt)
+            (place/rhousing-place getopt :interior nil nil
+              (prop :derived :position :hollow)
+              (misc/bevelled-cuboid
+                (prop :derived :size :hollow)
+                (prop :bevel :interior)))))
+        (when (fast :bosses)
+          (model/intersection
+            (rear-housing-exterior getopt)
+            (pair rhousing-mount-positive))))
+      (pair rhousing-mount-negative))))